nuke: don't ask if there's nothing to delete

+++ nuke.sh
@@ -5,14 +5,19 @@
 set -e
 git pull --rebase > /dev/null
 
-read -n 1 -r -p "Delete Codemagic application '${APP_NAME_SNAKE}'? (y/N) " YN
-echo
-if [[ "$YN" =~ ^[yY] ]]; then
-#    -H "Content-Type: application/json" \
-  curl  "https://api.codemagic.io/apps/${CODEMAGIC_APP_ID}" \
-    -H "x-auth-token: $(cat "$CM_API_TOKEN_PATH")" \
-    -X DELETE
+
+if [[ -n "$CODEMAGIC_APP_ID" ]]; then
+  read -n 1 -r -p "Delete Codemagic application '${APP_NAME_SNAKE}'? (y/N) " YN
   echo
+  if [[ "$YN" =~ ^[yY] ]]; then
+  #    -H "Content-Type: application/json" \
+    curl  "https://api.codemagic.io/apps/${CODEMAGIC_APP_ID}" \
+      -H "x-auth-token: $(cat "$CM_API_TOKEN_PATH")" \
+      -X DELETE
+    echo
+  fi
+else
+  echo "No Codemagic app Id found, skipping"
 fi
 
 # Google projects are heavy-weight entities, which take weeks to delete.
@@ -24,13 +29,30 @@ fi
 #  gcloud projects delete "$APP_ID_SLUG" --quiet
 #fi
 
-read -n 1 -r -p "Delete Sentry project '${SENTRY_PROJECT}'? (y/N) " YN
-echo
-if [[ "$YN" =~ ^[yY] ]]; then
-  curl "https://sentry.io/api/0/projects/${SENTRY_ORG}/${SENTRY_PROJECT}/" \
-   -H "Authorization: Bearer $(cat "$SENTRY_PROJECTS_ADMIN_TOKEN_PATH")" \
-   -X DELETE
-   echo
+
+if [[ -n "$SENTRY_DSN" ]]; then
+  read -n 1 -r -p "Delete Sentry project '${SENTRY_PROJECT}'? (y/N) " YN
+  echo
+  if [[ "$YN" =~ ^[yY] ]]; then
+    curl "https://sentry.io/api/0/projects/${SENTRY_ORG}/${SENTRY_PROJECT}/" \
+     -H "Authorization: Bearer $(cat "$SENTRY_PROJECTS_ADMIN_TOKEN_PATH")" \
+     -X DELETE
+     echo
+  fi
+  if ! cmp -s lib/main.dart.sentry lib/main.dart; then
+    read -n 1 -r -p "Push main.dart back to main.dart.sentry? (y/N) " YN
+    echo
+    if [[ "$YN" =~ ^[yY] ]]; then
+      cp -f lib/main.dart.sentry lib/main.dart.swap
+      cp -f lib/main.dart lib/main.dart.sentry
+      mv -f lib/main.dart.swap lib/main.dart
+      git add lib/main.dart*
+      git commit -m 'Handle main.dart'
+      git push
+    fi
+  fi
+else
+  echo "No Sentry DSN found, skipping"
 fi
 
 if [ -f shorebird.yaml ]; then
@@ -42,6 +64,8 @@ if [ -f shorebird.yaml ]; then
     echo "You need to do this manually at '${sbProjUrl}'"
     xdg-open "$sbProjUrl" > /dev/null
   fi
+else
+  echo "No Shorebird integration found, skipping"
 fi
 
 read -n 1 -r -p "Delete release tags from git? (y/N) " YN
@@ -49,17 +73,4 @@ echo
 if [[ "$YN" =~ ^[yY] ]]; then
   git tag -l | grep '+' | xargs -I {} git push origin :refs/tags/{}
   git tag -l | grep '+' | xargs git tag -d
-fi
-
-if [[ -n "$SENTRY_DSN" ]]; then
-  read -n 1 -r -p "Push main.dart back to main.dart.sentry? (y/N) " YN
-  echo
-  if [[ "$YN" =~ ^[yY] ]]; then
-    cp -f lib/main.dart.sentry lib/main.dart.swap
-    cp -f lib/main.dart lib/main.dart.sentry
-    mv -f lib/main.dart.swap lib/main.dart
-    git add lib/main.dart*
-    git commit -m 'Handle main.dart'
-    git push
-  fi
 fi
